Don't mirror parts of the directory structure of uClinux (and inherently trunk/) on the vendor branches.
Otherwise it's not easy to know where to put what when
importing new software.

Another question is what happens when we move something on the trunk,
say from user/ to lib/. To keep things consistent we'd have to move it
on the vendor branch as well for no good reason.

So let's just use a flat hierarchy from now on.